src/main/java/com/springHello/service/UserService.java
package com.springHello.service;

import com.springHello.bean.User;
import com.springHello.mapper.UserMapper;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

/**
 * Created by root on 1/26/17.
 */
@Service
@Transactional
public class UserService implements UserMapper {

    @Autowired
    private UserMapper userMapper;
    @Override
    public List<User> findAllUser() {
        return userMapper.findAllUser();
    }
}
